chore(drec-ui) separated logic for approve operator
