+++ apps/drec-ui/src/apps/user/data/index.ts
@@ -9,3 +9,4 @@ export * from './login';
 export * from './registerUser';
 export * from './fetching';
 export * from './handlers';
+export * from './approveOperator';
