+++ apps/drec-ui/src/apps/user/view/containers/blockchain/OrganizationBlockchainAddress/OrganizationBlockchainAddress.effects.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useUpdateBlockchainAddress } from 'apps/user/data';
+import { useApproveOperatorHandler, useUpdateBlockchainAddress } from 'apps/user/data';
 import { useOrganizationBlockchainAddressLogic } from 'apps/user/logic';
 import { useUserAppEnv } from '../../../context';
 
@@ -8,7 +8,8 @@ export const useOrganizationBlockchainAddressEffects = () => {
     const { registrationMessage } = useUserAppEnv();
 
     const api = useUpdateBlockchainAddress(registrationMessage, setIsUpdating);
+    const operatorApproval = useApproveOperatorHandler();
     const logic = useOrganizationBlockchainAddressLogic();
 
-    return { ...api, isUpdating, ...logic };
+    return { ...operatorApproval, ...api, isUpdating, ...logic };
 };
